Add back cache mount to package install (*Config).Install
